More moving
TS types will probably be moved to a Github Action
