+++ package.json
@@ -13,9 +13,9 @@
     "test": "mocha",
     "test:watch": "mocha --reporter min --watch",
     "coverage": "nyc --reporter=text mocha --reporter min",
+    "types": "tsc",
     "document": "jsdoc -c ./jsdoc.conf.json -R ./README.md && cp ./docs_statics/* ./docs/",
-    "document:delete": "rm -rf ./docs/*",
-    "types": "tsc"
+    "document:delete": "rm -rf ./docs/*"
   },
   "keywords": [
     "whatsapp",
